extract entity builder functions into dedicated file
https://coveord.atlassian.net/browse/KIT-411
